fix compilation with moving files
