+++ src/filter_strategy/AmplitudeFilter.cc
