+++ src/filter_strategy/CompositeFilter.cc
