+++ src/filter_strategy/ElevationNaNFilter.cc
