+++ src/filter_strategy/IFilterStrategy.h
