+++ src/filter_strategy/KeepFirstPingFilter.cc
