+++ src/filter_strategy/RatioAmpNoiseFilter.cc
