+++ src/ping_merge_strategy/IPingMergeStrategy.h
