+++ src/ping_merge_strategy/MeanMergeStrategy.cc
